requirements.txt
discord.py
simplejson
redis
requests
randomcolor
pillow
matplotlib
apnggif
switchcase